refactor: move all docker hub related code to DockerHubIntegration namespace
