+++ Doppler.CDHelper.Test/DockerHubHandlerTest.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Net;
 using System.Net.Http.Json;
 using System.Threading.Tasks;
-using Doppler.CDHelper.Controllers;
+using Doppler.CDHelper.DockerHubIntegration;
 using Doppler.CDHelper.SwarmClient;
 using Doppler.CDHelper.SwarmServiceSelection;
 using Flurl.Http.Testing;
