+++ Doppler.CDHelper/DockerHubIntegration/DockerHubHookData.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 
-namespace Doppler.CDHelper
+namespace Doppler.CDHelper.DockerHubIntegration
 {
     [SuppressMessage("Naming", "CA1707:Identifiers should not contain underscores", Justification = "Allow properties with underscore because it represents a JSON")]
     [SuppressMessage("Style", "IDE1006:Naming Styles", Justification = "Allow properties with camelCase because it represents a JSON")]
