+++ Doppler.CDHelper/SwarmServiceSelection/ISwarmServiceSelector.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Doppler.CDHelper.DockerHubIntegration;
 using Doppler.CDHelper.SwarmClient;
 
 namespace Doppler.CDHelper.SwarmServiceSelection
